+++ src/api/authlete_api.ts
@@ -20,14 +20,24 @@ import { AuthorizationRequest } from '../dto/authorization_request.ts';
 import { AuthorizationResponse } from '../dto/authorization_response.ts';
 import { Client } from '../dto/client.ts';
 import { ClientListResponse } from '../dto/client_list_response.ts';
+import { IntrospectionRequest } from '../dto/introspection_request.ts';
+import { IntrospectionResponse } from '../dto/introspection_response.ts';
+import { RevocationRequest } from '../dto/revocation_request.ts';
+import { RevocationResponse } from '../dto/revocation_response.ts';
 import { Service } from '../dto/service.ts';
 import { ServiceListResponse } from '../dto/service_list_response.ts';
+import { StandardIntrospectionRequest } from '../dto/standard_introspection_request.ts';
+import { StandardIntrospectionResponse } from '../dto/standard_introspection_response.ts';
 import { TokenFailRequest } from '../dto/token_fail_request.ts';
 import { TokenFailResponse } from '../dto/token_fail_response.ts';
 import { TokenIssueRequest } from '../dto/token_issue_request.ts';
 import { TokenIssueResponse } from '../dto/token_issue_response.ts';
 import { TokenRequest } from '../dto/token_request.ts';
 import { TokenResponse } from '../dto/token_response.ts';
+import { UserInfoIssueRequest } from '../dto/user_info_issue_request.ts';
+import { UserInfoIssueResponse } from '../dto/user_info_issue_response.ts';
+import { UserInfoRequest } from '../dto/user_info_request.ts';
+import { UserInfoResponse } from '../dto/user_info_response.ts';
 
 
 /**
@@ -38,7 +48,8 @@ export interface AuthleteApi
     /**
      * Call `/auth/authorization` API.
      *
-     * @param request - A request.
+     * @param request
+     *         Request parameters passed to the API.
      */
     authorization(request: AuthorizationRequest): Promise<AuthorizationResponse>;
 
@@ -46,7 +57,8 @@ export interface AuthleteApi
     /**
      * Call `/auth/authorization/issue` API.
      *
-     * @param request - A request.
+     * @param request
+     *         Request parameters passed to the API.
      */
     authorizationIssue(request: AuthorizationIssueRequest): Promise<AuthorizationIssueResponse>;
 
@@ -54,7 +66,8 @@ export interface AuthleteApi
     /**
      * Call `/auth/authorization/fail` API.
      *
-     * @param request - A request.
+     * @param request
+     *         Request parameters passed to the API.
      */
     authorizationFail(request: AuthorizationFailRequest): Promise<AuthorizationFailResponse>;
 
@@ -62,15 +75,17 @@ export interface AuthleteApi
     /**
      * Call `/auth/token` API.
      *
-     * @param request - A request.
+     * @param request
+     *         Request parameters passed to the API.
      */
-    token(request: TokenRequest):Promise<TokenResponse>;
+    token(request: TokenRequest): Promise<TokenResponse>;
 
 
     /**
      * Call `/auth/token/issue` API.
      *
-     * @param request - A request.
+     * @param request
+     *         Request parameters passed to the API.
      */
     tokenIssue(request: TokenIssueRequest): Promise<TokenIssueResponse>;
 
@@ -78,15 +93,62 @@ export interface AuthleteApi
     /**
      * Call `/auth/token/fail` API.
      *
-     * @param request - A request.
+     * @param request
+     *         Request parameters passed to the API.
      */
     tokenFail(request: TokenFailRequest): Promise<TokenFailResponse>;
 
 
     /**
+     * Call Authlete `/auth/revocation` API.
+     *
+     * @param request
+     *         Request parameters passed to the API.
+     */
+    revocation(request: RevocationRequest): Promise<RevocationResponse>;
+
+
+    /**
+     * Call Authlete `/auth/userinfo` API.
+     *
+     * @param request
+     *         Request parameters passed to the API.
+     */
+    userInfo(request: UserInfoRequest): Promise<UserInfoResponse>;
+
+
+    /**
+     * Call Authlete `/auth/userinfo/issue` API.
+     *
+     * @param request
+     *         Request parameters passed to the API.
+     */
+    userInfoIssue(request: UserInfoIssueRequest): Promise<UserInfoIssueResponse>;
+
+
+    /**
+     * Call Authlete `/auth/introspection` API.
+     *
+     * @param request
+     *         Request parameters passed to the API.
+     */
+    introspection(request: IntrospectionRequest): Promise<IntrospectionResponse>;
+
+
+    /**
+     * Call Authlete `/auth/introspection/standard` API.
+     *
+     * @param request
+     *         Request parameters passed to the API.
+     */
+    standardIntrospection(request: StandardIntrospectionRequest): Promise<StandardIntrospectionResponse>;
+
+
+    /**
      * Get information about a service.
      *
-     * @param apiKey - The API key of a service.
+     * @param apiKey
+     *         The API key of a service.
      */
     getService(apiKey: number): Promise<Service>;
 
@@ -94,11 +156,13 @@ export interface AuthleteApi
     /**
      * Get list of services.
      *
-     * @param start - The start index (inclusive) of the result set.
-     *                The default value is 0. Must not be a negative number.
+     * @param start
+     *         The start index (inclusive) of the result set. The default
+     *         value is 0. Must not be a negative number.
      *
-     * @param end - The end index (exclusive) of the result set. The default
-     *              value is 5. Must not be a negative number.
+     * @param end
+     *         The end index (exclusive) of the result set. The default
+     *         value is 5. Must not be a negative number.
      */
     getServiceList(start?: number, end?: number): Promise<ServiceListResponse>
 
@@ -106,7 +170,8 @@ export interface AuthleteApi
     /**
      * Create a service.
      *
-     * @param service A `Service` object representing a new service.
+     * @param service
+     *         A service to create.
      */
     createService(service: Service): Promise<Service>
 
@@ -114,7 +179,8 @@ export interface AuthleteApi
     /**
      * Update a service.
      *
-     * @param service A `Service` object representing an updated service.
+     * @param service
+     *         A service to update.
      */
     updateService(service: Service): Promise<Service>
 
@@ -122,15 +188,61 @@ export interface AuthleteApi
     /**
      * Delete a service.
      *
-     * @param apiKey - The API key of a service.
+     * @param apiKey
+     *         The API key of a service.
      */
     deleteService(apiKey: number): Promise<void>
 
 
     /**
+     * Get the JWK Set of a service.
+     *
+     * This methods returns the value of the service's `jwks` property
+     * or the value provided at the service's `jwksUri` location. If
+     * both properties have been set, the value of the `jwks` property
+     * is returned in preference to the other one.
+     *
+     * For more details, see [RFC 7517: JSON Web Key (JWK)](
+     * https://tools.ietf.org/html/rfc7517).
+     *
+     * @param pretty
+     *         `true` to get the JSON in pretty format.
+     *
+     * @param includePrivateKeys
+     *         `true` to keep private keys in the JSON. `false` to remove
+     *         private keys.
+     *
+     * @returns JSON representation of the JWK Set of the service. `null`
+     *          is returned when the service has registered neither content
+     *          or URI of its JWK Set.
+     */
+    getServiceJwks(pretty?: boolean, includePrivateKeys?: boolean): Promise<string | null>;
+
+
+    /**
+     * Get the configuration of the service in JSON format that complies
+     * with [OpenID Connect Discovery 1.0](
+     * http://openid.net/specs/openid-connect-discovery-1_0.html).
+     *
+     * The value returned from this method can be used as the response
+     * body from `/.well-known/openid-configuration`. See "[4. Obtaining
+     * OpenID Provider Configuration Information](
+     * http://openid.net/specs/openid-connect-discovery-1_0.html#ProviderConfig)"
+     * in OpenID Connect Discovery 1.0 for details.
+     *
+     * @param pretty
+     *         `true` to get the JSON in pretty format.
+     *
+     * @returns The configuration of the service in JSON format.
+     */
+    getServiceConfiguration(pretty?: boolean): Promise<string>;
+
+
+    /**
      * Get information about a client application.
      *
-     * @param clientId - The client ID of a client application.
+     * @param clientId
+     *         The client ID of a client application.
      */
     getClient(clientId: number): Promise<Client>
 
@@ -138,17 +250,19 @@ export interface AuthleteApi
     /**
      * Get list of client applications.
      *
-     * @param developer - The developer of client applications. If this
-     *                    parameter is specified, client applications
-     *                    of the specified developer are returned.
-     *                    Otherwise, all the client applications that
-     *                    belong to the service are returned.
+     * @param developer
+     *         The developer of client applications. If this parameter
+     *         is specified, client applications of the specified developer
+     *         are returned. Otherwise, all the client applications that
+     *         belong to the service are returned.
      *
-     * @param start - The start index (inclusive) of the result set.
-     *                The default value is 0. Must not be a negative number.
+     * @param start
+     *         The start index (inclusive) of the result set. The default
+     *         value is 0. Must not be a negative number.
      *
-     * @param end - The end index (exclusive) of the result set. The default
-     *              value is 5. Must not be a negative number.
+     * @param end
+     *         The end index (exclusive) of the result set. The default
+     *         value is 5. Must not be a negative number.
      */
     getClientList(developer?: string, start?: number, end?: number): Promise<ClientListResponse>;
 
@@ -156,7 +270,8 @@ export interface AuthleteApi
     /**
      * Create a client application.
      *
-     * @param client A `Client` object representing a new client application.
+     * @param client
+     *         A client application to create.
      */
     createClient(client: Client): Promise<Client>;
 
@@ -164,7 +279,8 @@ export interface AuthleteApi
     /**
      * Update a client application.
      *
-     * @param client A `Client` object representing an updated client application.
+     * @param client
+     *         A client application to update.
      */
     updateClient(client: Client): Promise<Client>;
 
@@ -172,7 +288,8 @@ export interface AuthleteApi
     /**
      * Delete a client application.
      *
-     * @param clientId - The client ID of a client application.
+     * @param clientId
+     *         The client ID of a client application.
      */
     deleteClient(clientId: number): Promise<void>;
 }