+++ CHANGES.md
@@ -1,6 +1,20 @@
 CHANGES
 =======
 
+1.1.0 (2020-07-02)
+------------------
+
+- Add support for the following APIs.
+
+  - /api/auth/introspection API
+  - /api/auth/introspection/standard API
+  - /api/auth/revocation API
+  - /api/auth/userinfo API
+  - /api/auth/userinfo/issue API
+  - /api/service/configuration API
+  - /api/service/jwks API
+
+
 1.0.0 (2020-06-28)
 ------------------